made 06_04 version 2, and v1 full completed
In the process of making 06_05

Reverted 06_02 into "state" where enemies are not required to be killed. Just press button to unblock capture area

+++ AreaRegionRelated/Subs/ObjectDestroyingAreaRegion/ObjectDestroyingAreaRegion.gd
@@ -20,8 +20,9 @@ func _ready():
 
 func _on_region__body_entered_in_area_o(body):
 	if body.get("is_class_type_base_object"):
-		if !is_hidden_and_silent:
-			AudioManager.helper__play_sound_effect__2d(StoreOfAudio.AudioIds.SFX_AreaRegion_ObjectDestroyed, body.global_position, 1.5, null)
-		
-		#body.queue_free()
-		body.destroy_self_caused_by_destroying_area_region(self)
+		if !body.ignore_object_destroying_region:
+			if !is_hidden_and_silent:
+				AudioManager.helper__play_sound_effect__2d(StoreOfAudio.AudioIds.SFX_AreaRegion_ObjectDestroyed, body.global_position, 1.5, null)
+			
+			#body.queue_free()
+			body.destroy_self_caused_by_destroying_area_region(self)
